test_vao_attribs
